Generalize Filters Slideover (#5158)
* Adds `FiltersSlideover`

* Upgrade `FacilityFilter` to use `FiltersSlideover`

* rename: `PatientFilterV2` -> `PatientFilter`

* minor fix in `FiltersSlideover`

* Upgrade `PatientFilter` to use `FiltersSlideover`

* minor fix in `FiltersSlideover`

* Upgrade `AssetFilter` to use `FiltersSlideover`

* Upgrade `SampleFilters` to use `FiltersSlideover`

* Upgrade External Result filter to use `FiltersSlideover`

* Upgrade `UserFilter` to use `FiltersSlideover`

* Upgrade Resource and Shifting Filters

* Remove no longer used `FilterButtons` component

* minor ui fix

* translations

* fix `FiltersSlideover` max-w

* generalize Advanced Filters buton

* fix filters not clearing

* fix asset location state management

---------
