--- src/Components/Common/AdvancedFilterButton.tsx
@@ -1,47 +1,0 @@
-import React from "react";
-
-const AdvancedFilterButton = ({ setShowFilters }: any) => {
-  return (
-    <div>
-      <div className="flex items-start">
-        <button
-          className="btn btn-primary-ghost w-full"
-          onClick={() => setShowFilters(true)}
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            className="fill-current w-4 h-4 mr-2"
-          >
-            <line x1="8" y1="6" x2="21" y2="6"></line>
-            <line x1="8" y1="12" x2="21" y2="12">
-              {" "}
-            </line>
-            <line x1="8" y1="18" x2="21" y2="18">
-              {" "}
-            </line>
-            <line x1="3" y1="6" x2="3.01" y2="6">
-              {" "}
-            </line>
-            <line x1="3" y1="12" x2="3.01" y2="12">
-              {" "}
-            </line>
-            <line x1="3" y1="18" x2="3.01" y2="18">
-              {" "}
-            </line>
-          </svg>
-          <span>Advanced Filters</span>
-        </button>
-      </div>
-    </div>
-  );
-};
-
-export default AdvancedFilterButton;
