+++ src/Components/Facility/FacilityFilter/LocalBodySelect.tsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
 import { getLocalbodyByDistrict } from "../../../Redux/actions";
 import { LegacyAutoCompleteAsyncField } from "../../Common/HelperInputFields";
@@ -27,6 +28,7 @@ function LocalBodySelect(props: LocalBodySelectProps) {
     margin,
     district,
   } = props;
+  const { t } = useTranslation();
   const [loadBodyLoading, isLocalBodyLoading] = useState(false);
   const [selectedObject, setSelectedObject] = useState(null);
   const [hasSearchText, setHasSearchText] = useState(false);
@@ -91,11 +93,11 @@ function LocalBodySelect(props: LocalBodySelectProps) {
       onSearch={handleSearch}
       onChange={(e: any, selected: any) => handleValueChange(selected)}
       loading={loadBodyLoading}
-      placeholder="Enter local body name"
+      placeholder={t("select_local_body")}
       noOptionsText={
         hasSearchText
           ? "No local body found, please try again"
-          : "Start typing to begin search"
+          : t("type_to_search")
       }
       renderOption={(option: any) => <div>{option.name}</div>}
       getOptionSelected={(option: any, value: any) => option.id === value.id}
