+++ src/Components/Patient/PatientHome.tsx
@@ -651,17 +651,16 @@ export const PatientHome = (props: any) => {
                     </div>
                   </div>
                 )}
-                {patientData.is_vaccinated &&
-                  patientData.last_vaccinated_date && (
-                    <div className="sm:col-span-1">
-                      <div className="text-sm leading-5 font-semibold text-zinc-400">
-                        Last Vaccinated on
-                      </div>
-                      <div className="mt-1 text-sm leading-5 font-medium">
-                        {formatDate(patientData.last_vaccinated_date)}
-                      </div>
+                {patientData.is_vaccinated && patientData.last_vaccinated_date && (
+                  <div className="sm:col-span-1">
+                    <div className="text-sm leading-5 font-semibold text-zinc-400">
+                      Last Vaccinated on
                     </div>
-                  )}
+                    <div className="mt-1 text-sm leading-5 font-medium">
+                      {formatDate(patientData.last_vaccinated_date)}
+                    </div>
+                  </div>
+                )}
                 {patientData.countries_travelled &&
                   !!patientData.countries_travelled.length && (
                     <div className="sm:col-span-1">
