+++ src/Components/Shifting/ListView.tsx
@@ -8,7 +8,6 @@ import {
   completeTransfer,
   downloadShiftRequests,
 } from "../../Redux/actions";
-import { make as SlideOver } from "../Common/SlideOver.gen";
 import ListFilter from "./ListFilter";
 import { formatFilter } from "./Commons";
 import { formatDate } from "../../Utils/utils";
@@ -340,11 +339,7 @@ export default function ListView() {
           </div>
         )}
       </div>
-      <SlideOver {...advancedFilter}>
-        <div className="bg-white min-h-screen p-4">
-          <ListFilter showShiftingStatus={true} {...advancedFilter} />
-        </div>
-      </SlideOver>
+      <ListFilter showShiftingStatus={true} {...advancedFilter} />
     </Page>
   );
 }
